Progress on test refactoring
