+++ sparse/unit_test/Test_Sparse_par_ilut.hpp
@@ -29,6 +29,8 @@
 #include "KokkosSparse_LUPrec.hpp"
 #include "KokkosSparse_SortCrs.hpp"
 
+#include "Test_vector_fixtures.hpp"
+
 #include <gtest/gtest.h>
 
 using namespace KokkosSparse;
@@ -54,69 +56,6 @@ struct TolMeta<float> {
 
 template <typename scalar_t, typename lno_t, typename size_type,
           typename device>
-std::vector<std::vector<scalar_t>> decompress_matrix(
-    Kokkos::View<size_type*, device>& row_map,
-    Kokkos::View<lno_t*, device>& entries,
-    Kokkos::View<scalar_t*, device>& values) {
-  const size_type nrows = row_map.size() - 1;
-  std::vector<std::vector<scalar_t>> result;
-  result.resize(nrows);
-  for (auto& row : result) {
-    row.resize(nrows, 0.0);
-  }
-
-  auto hrow_map = Kokkos::create_mirror_view(row_map);
-  auto hentries = Kokkos::create_mirror_view(entries);
-  auto hvalues  = Kokkos::create_mirror_view(values);
-  Kokkos::deep_copy(hrow_map, row_map);
-  Kokkos::deep_copy(hentries, entries);
-  Kokkos::deep_copy(hvalues, values);
-
-  for (size_type row_idx = 0; row_idx < nrows; ++row_idx) {
-    const size_type row_nnz_begin = hrow_map(row_idx);
-    const size_type row_nnz_end   = hrow_map(row_idx + 1);
-    for (size_type row_nnz = row_nnz_begin; row_nnz < row_nnz_end; ++row_nnz) {
-      const lno_t col_idx      = hentries(row_nnz);
-      const scalar_t value     = hvalues(row_nnz);
-      result[row_idx][col_idx] = value;
-    }
-  }
-
-  return result;
-}
-
-template <typename scalar_t, typename lno_t, typename size_type,
-          typename device>
-void check_matrix(const std::string& name,
-                  Kokkos::View<size_type*, device>& row_map,
-                  Kokkos::View<lno_t*, device>& entries,
-                  Kokkos::View<scalar_t*, device>& values,
-                  const std::vector<std::vector<scalar_t>>& expected) {
-  const auto decompressed_mtx = decompress_matrix(row_map, entries, values);
-
-  const size_type nrows = row_map.size() - 1;
-  for (size_type row_idx = 0; row_idx < nrows; ++row_idx) {
-    for (size_type col_idx = 0; col_idx < nrows; ++col_idx) {
-      EXPECT_NEAR(expected[row_idx][col_idx],
-                  decompressed_mtx[row_idx][col_idx], 0.01)
-          << "Failed check is: " << name << "[" << row_idx << "][" << col_idx
-          << "]";
-    }
-  }
-}
-
-template <typename scalar_t>
-void print_matrix(const std::vector<std::vector<scalar_t>>& matrix) {
-  for (const auto& row : matrix) {
-    for (const auto& item : row) {
-      std::printf("%.2f ", item);
-    }
-    std::cout << std::endl;
-  }
-}
-
-template <typename scalar_t, typename lno_t, typename size_type,
-          typename device>
 void run_test_par_ilut() {
   using RowMapType   = Kokkos::View<size_type*, device>;
   using EntriesType  = Kokkos::View<lno_t*, device>;
@@ -126,52 +65,20 @@ void run_test_par_ilut() {
       typename device::memory_space, typename device::memory_space>;
 
   // Simple test fixture A
-  std::vector<std::vector<scalar_t>> A = {{1., 6., 4., 7.},
-                                          {2., -5., 0., 8.},
-                                          {0.5, -3., 6., 0.},
-                                          {0.2, -0.5, -9., 0.}};
-
-  const scalar_t ZERO = scalar_t(0);
-
-  const size_type nrows = A.size();
-
-  // Count A nnz's
-  size_type nnz = 0;
-  for (size_type row_idx = 0; row_idx < nrows; ++row_idx) {
-    for (size_type col_idx = 0; col_idx < nrows; ++col_idx) {
-      if (A[row_idx][col_idx] != ZERO) {
-        ++nnz;
-      }
-    }
-  }
+  std::vector<std::vector<scalar_t>> A = {{1.,   6.,   4.,  7.},
+                                          {2.,  -5.,   0.,  8.},
+                                          {0.5, -3.,   6.,  0.},
+                                          {0.2, -0.5, -9.,  0.}};
 
   // Allocate device CRS views for A
-  RowMapType row_map("row_map", nrows + 1);
-  EntriesType entries("entries", nnz);
-  ValuesType values("values", nnz);
+  RowMapType row_map("row_map");
+  EntriesType entries("entries");
+  ValuesType values("values");
 
-  // Create host mirror views for CRS A
-  auto hrow_map = Kokkos::create_mirror_view(row_map);
-  auto hentries = Kokkos::create_mirror_view(entries);
-  auto hvalues  = Kokkos::create_mirror_view(values);
+  compress_matrix(row_map, entries, values, A);
 
-  // Compress A into CRS (host views)
-  size_type curr_nnz = 0;
-  for (size_type row_idx = 0; row_idx < nrows; ++row_idx) {
-    for (size_type col_idx = 0; col_idx < nrows; ++col_idx) {
-      if (A[row_idx][col_idx] != ZERO) {
-        hentries(curr_nnz) = col_idx;
-        hvalues(curr_nnz)  = A[row_idx][col_idx];
-        ++curr_nnz;
-      }
-      hrow_map(row_idx + 1) = curr_nnz;
-    }
-  }
-
-  // Copy host A CRS views to device A CRS views
-  Kokkos::deep_copy(row_map, hrow_map);
-  Kokkos::deep_copy(entries, hentries);
-  Kokkos::deep_copy(values, hvalues);
+  const size_type nrows = A.size();
+  const size_type nnz   = values.extent(0);
 
   // Make kernel handle
   KernelHandle kh;
