+++ sparse/unit_test/Test_vector_fixtures.hpp
@@ -1,0 +1,194 @@
+//@HEADER
+// ************************************************************************
+//
+//                        Kokkos v. 4.0
+//       Copyright (2022) National Technology & Engineering
+//               Solutions of Sandia, LLC (NTESS).
+//
+// Under the terms of Contract DE-NA0003525 with NTESS,
+// the U.S. Government retains certain rights in this software.
+//
+// Part of Kokkos, under the Apache License v2.0 with LLVM Exceptions.
+// See https://kokkos.org/LICENSE for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//@HEADER
+
+#ifndef _TEST_VECTOR_FIXTURES_HPP
+#define _TEST_VECTOR_FIXTURES_HPP
+
+#include <Kokkos_Core.hpp>
+
+#include <vector>
+
+/**
+ * API for working with 2D vectors of small matrices for testing.
+ */
+
+namespace Test {
+
+template <typename scalar_t, typename lno_t, typename size_type,
+          typename device>
+void compress_matrix(
+    Kokkos::View<size_type*, device>& row_map,
+    Kokkos::View<lno_t*, device>& entries,
+    Kokkos::View<scalar_t*, device>& values,
+    const std::vector<std::vector<scalar_t>>& fixture)
+{
+  const scalar_t ZERO = scalar_t(0);
+
+  const size_type nrows = fixture.size();
+
+  // Count fixture nnz's
+  size_type nnz = 0;
+  for (size_type row_idx = 0; row_idx < nrows; ++row_idx) {
+    for (size_type col_idx = 0; col_idx < nrows; ++col_idx) {
+      if (fixture[row_idx][col_idx] != ZERO) {
+        ++nnz;
+      }
+    }
+  }
+
+  // Allocate device CRS views
+  Kokkos::resize(row_map, nrows + 1);
+  Kokkos::resize(entries, nnz);
+  Kokkos::resize(values, nnz);
+
+  // Create host mirror views for CRS
+  auto hrow_map = Kokkos::create_mirror_view(row_map);
+  auto hentries = Kokkos::create_mirror_view(entries);
+  auto hvalues  = Kokkos::create_mirror_view(values);
+
+  // Compress into CRS (host views)
+  size_type curr_nnz = 0;
+  for (size_type row_idx = 0; row_idx < nrows; ++row_idx) {
+    for (size_type col_idx = 0; col_idx < nrows; ++col_idx) {
+      if (fixture[row_idx][col_idx] != ZERO) {
+        hentries(curr_nnz) = col_idx
+        hvalues(curr_nnz)  = fixture[row_idx][col_idx];
+        ++curr_nnz;
+      }
+      hrow_map(row_idx + 1) = curr_nnz;
+    }
+  }
+
+  // Copy host CRS views to device CRS views
+  Kokkos::deep_copy(row_map, hrow_map);
+  Kokkos::deep_copy(entries, hentries);
+  Kokkos::deep_copy(values, hvalues);
+}
+
+template <typename scalar_t, typename lno_t, typename size_type,
+          typename device>
+std::vector<std::vector<scalar_t>> decompress_matrix(
+    Kokkos::View<size_type*, device>& row_map,
+    Kokkos::View<lno_t*, device>& entries,
+    Kokkos::View<scalar_t*, device>& values)
+{
+  const scalar_t ZERO = scalar_t(0);
+
+  const size_type nrows = row_map.size() - 1;
+  std::vector<std::vector<scalar_t>> result;
+  result.resize(nrows);
+  for (auto& row : result) {
+    row.resize(nrows, ZERO);
+  }
+
+  auto hrow_map = Kokkos::create_mirror_view(row_map);
+  auto hentries = Kokkos::create_mirror_view(entries);
+  auto hvalues  = Kokkos::create_mirror_view(values);
+  Kokkos::deep_copy(hrow_map, row_map);
+  Kokkos::deep_copy(hentries, entries);
+  Kokkos::deep_copy(hvalues, values);
+
+  for (size_type row_idx = 0; row_idx < nrows; ++row_idx) {
+    const size_type row_nnz_begin = hrow_map(row_idx);
+    const size_type row_nnz_end   = hrow_map(row_idx + 1);
+    for (size_type row_nnz = row_nnz_begin; row_nnz < row_nnz_end; ++row_nnz) {
+      const lno_t col_idx      = hentries(row_nnz);
+      const scalar_t value     = hvalues(row_nnz);
+      result[row_idx][col_idx] = value;
+    }
+  }
+
+  return result;
+}
+
+template <typename scalar_t, typename lno_t, typename size_type,
+          typename device>
+std::vector<std::vector<typename View3::non_const_value_type>> decompress_matrix(
+  Kokkos::View<size_type*, device>& row_map,
+  Kokkos::View<lno_t*, device>& entries,
+  Kokkos::View<scalar_t*, device>& values,
+  const int block_size)
+{
+  const scalar_t ZERO = scalar_t(0);
+
+  const size_type nbrows   = row_map.extent(0) - 1;
+  const size_type nrows    = nbrows * block_size;
+  const size_type block_items = block_size * block_size;
+  std::vector<std::vector<scalar_t>> result;
+  result.resize(nrows);
+  for (auto& row : result) {
+    row.resize(nrows, ZERO);
+  }
+
+  auto hrow_map = Kokkos::create_mirror_view(row_map);
+  auto hentries = Kokkos::create_mirror_view(entries);
+  auto hvalues  = Kokkos::create_mirror_view(values);
+  Kokkos::deep_copy(hrow_map, row_map);
+  Kokkos::deep_copy(hentries, entries);
+  Kokkos::deep_copy(hvalues, values);
+
+  for (size_type row_idx = 0; row_idx < nbrows; ++row_idx) {
+    const size_type row_nnz_begin = hrow_map(row_idx);
+    const size_type row_nnz_end   = hrow_map(row_idx + 1);
+    for (size_type row_nnz = row_nnz_begin; row_nnz < row_nnz_end; ++row_nnz) {
+      const lno_t col_idx = hentries(row_nnz);
+      for (size_type i = 0; i < block_size; ++i) {
+        const size_type unc_row_idx = row_idx*block_size + i;
+        for (size_type j = 0; j < block_size; ++j) {
+          const size_type unc_col_idx = col_idx*block_size + j;
+          result[unc_row_idx][unc_col_idx] = hvalues(row_nnz*block_items + i*block_size + j);
+        }
+      }
+    }
+  }
+
+  return result;
+}
+
+
+template <typename scalar_t, typename lno_t, typename size_type,
+          typename device>
+void check_matrix(const std::string& name,
+                  Kokkos::View<size_type*, device>& row_map,
+                  Kokkos::View<lno_t*, device>& entries,
+                  Kokkos::View<scalar_t*, device>& values,
+                  const std::vector<std::vector<scalar_t>>& expected) {
+  const auto decompressed_mtx = decompress_matrix(row_map, entries, values);
+
+  const size_type nrows = row_map.size() - 1;
+  for (size_type row_idx = 0; row_idx < nrows; ++row_idx) {
+    for (size_type col_idx = 0; col_idx < nrows; ++col_idx) {
+      EXPECT_NEAR(expected[row_idx][col_idx],
+                  decompressed_mtx[row_idx][col_idx], 0.01)
+          << "Failed check is: " << name << "[" << row_idx << "][" << col_idx
+          << "]";
+    }
+  }
+}
+
+template <typename scalar_t>
+void print_matrix(const std::vector<std::vector<scalar_t>>& matrix) {
+  for (const auto& row : matrix) {
+    for (const auto& item : row) {
+      std::printf("%.2f ", item);
+    }
+    std::cout << std::endl;
+  }
+}
+
+} // namespace Test
+
+#endif  // _TEST_VECTOR_FIXTURES_HPP
